python/requirements.txt
kafka-python
spotipy
tweepy
